% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_var.R
\name{df_var}
\alias{df_var}
\title{df_var}
\usage{
df_var(df)
}
\arguments{
\item{df}{The input dataframe}
}
\value{
It returns a vector of variance corresponds to each numeric column in the dataframe.
}
\description{
This function calculates the variance of each column in a dataframe, it requires the dataframe to
contain only numeric columns.
}
\examples{
a <- c(23,45,34,231)
b <- c(0.3,3,22,0.003)
df <-  data.frame(a,b)
df_var(df)

}
